refactor: move completeOrderAsSuccess to util per review
